Broken but getting there
Added in controls. They don't work. This is largely due to my hacky way of finding what cubes to rotate.
The rotation isn't 100% accurate so they'll never perfectly equal their starting positions again.

To fix this I need to come up with an internal representation of the object to apply the moves to.
This can then update the model.